getting ini file's values from command center
