+++ DataBaze/DataBaze.vcxproj
@@ -193,6 +193,7 @@
     <ClInclude Include="Source\Public\Components\DBListBox.h" />
     <ClInclude Include="Source\Public\Components\DBSettingsComp.h" />
     <ClInclude Include="Source\Public\Core\DBCommandCenter.h" />
+    <ClInclude Include="Source\Public\DataTypes\DBSettingsTypes.h" />
     <ClInclude Include="Source\Public\DataTypes\Size2D.h" />
     <ClInclude Include="Source\Public\DBDataTypes.h" />
     <ClInclude Include="Source\Public\Components\DBDataManager.h" />
@@ -224,6 +225,7 @@
     <ClCompile Include="Source\Private\Components\DBWindowsManager.cpp" />
     <ClCompile Include="Source\Private\Core\DBCommandCenter.cpp" />
     <ClCompile Include="Source\Private\Core\DBInstance.cpp" />
+    <ClCompile Include="Source\Private\DataTypes\DBSettingsTypes.cpp" />
     <ClCompile Include="Source\Private\DBDataTypes.cpp" />
     <ClCompile Include="Source\Private\DBInterface.cpp" />
     <ClCompile Include="Source\Private\DBJsonUtils.cpp" />
