+++ DataBaze/Source/Private/DataTypes/DBSettingsTypes.cpp
@@ -1,0 +1,5 @@
+#include "DBSettingsTypes.h"
+
+// Initialize constants
+const DBIniItem DBIniItem::AutoFill_Delay = DBIniItem("Autofill", "FillDelay", "10.0");
+const DBIniItem DBIniItem::AutoFill_Other = DBIniItem("Autofill", "OtherParam", "20.0");
